GenAi/YT_Transcriber/requirements.txt
streamlit
youtube-transcript-api
google-generativeai
fpdf